Move graphql tags to components
